move docs/examples/datastructures/pagination.py > docs/examples/pagination.py
