+++ docs/examples/tests/pagination/test_using_offset_pagination.py
@@ -1,4 +1,4 @@
-from examples.datastructures.pagination.using_offset_pagination import app
+from examples.pagination.using_offset_pagination import app
 
 from starlite.status_codes import HTTP_200_OK
 from starlite.testing import TestClient
